Move Payment into separate package payment
