+++ sdk/src/main/java/io/singularitynet/sdk/client/BaseServiceClient.java
@@ -7,7 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import io.singularitynet.sdk.daemon.DaemonConnection;
-import io.singularitynet.sdk.daemon.Payment;
+import io.singularitynet.sdk.payment.Payment;
 import io.singularitynet.sdk.registry.MetadataProvider;
 import io.singularitynet.sdk.mpe.PaymentChannelStateProvider;
