+++ sdk/src/test/java/io/singularitynet/sdk/daemon/DaemonMock.java
@@ -13,6 +13,8 @@ import com.google.protobuf.ByteString;
 
 import io.singularitynet.daemon.escrow.*;
 import io.singularitynet.sdk.common.Utils;
+import io.singularitynet.sdk.payment.Payment;
+import io.singularitynet.sdk.payment.PaymentSerializer;
 import io.singularitynet.sdk.mpe.PaymentChannel;
 
 public class DaemonMock extends PaymentChannelStateServiceGrpc.PaymentChannelStateServiceImplBase
